private route bug fix
